refactor: fix field types
- change double fields to bigDecimal
- create PaymentStatus enum to avoid using simple String fields

+++ src/main/java/com/fiap/cheffyorderservice/application/ports/in/records/PlaceOrderCommandRecord.java
@@ -1,9 +1,10 @@
 package com.fiap.cheffyorderservice.application.ports.in.records;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PlaceOrderCommandRecord(
         UUID orderId,
-        Double totalAmount
+        BigDecimal totalAmount
 ) {
 }
